Sun, Jun 19, 2022, 3:55 PM +00:00

+++ pset1/credit.c
@@ -40,7 +40,7 @@ int get_digits(long int n, int pos)
 
 int len_number(long int n)
 {
-    int res = 1;
+    long int res = 1;
     int i = 0;
     while (res >= 0)
     {
